+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\core\accessors\setKeyValue
 - Adds oihana\core\date\formatDateTime
 - Adds oihana\core\reflections\getFunctionInfo
+- Adds oihana\core\strings\formatFromDocument
 - Adds oihana\core\strings\isRegexp
 
 ### Removed